refactor: optimize batch execution security and readability
- Replaced `Math.random()` with `randomInt()` from `node:crypto` in `batchExecuteMany` to ensure a cryptographically secure random distribution for the backoff jitter, preventing clustering in distributed architectures.
- Refined the warning message in `parsePositiveInt` to explicitly state when invalid config inputs are skipped and fallbacks are used.
- Extracted `BatchExecutionError` into a dedicated `src/utils/errors.ts` module to adhere to SRP and keep `database/utils.ts` focused strictly on functional database utilities.
